requirements.txt
flask==3.1.0
flask-restx==1.3.0
flask_cors
pymongo==4.10.0
flask_cors
pytest-selenium 
selenium 
webdriver-manager